Nagubandi Varshith/requirements.txt
streamlit
requests
streamlit-webrtc
openai-whisper
pytesseract
pillow
torch
opencv-python